Oranges vs. Grapefruit
The task of separating oranges and grapefruit is fairly obvious to a human, but even with manual observation there is still a bit of error. This dataset takes the color, weight, and diameter of an "average" orange and grapefruit and generates a larger dataset containing a wide variety of values and are "oranges" and "grapefruit".